backend/requirements.txt
Flask==2.3.2
pymongo==4.6.0
flask-cors==4.0.0
python-jwt==4.1.0
bcrypt==4.0.1
python-dotenv==1.0.0